Moved /core/server/data/meta to /core/frontend/meta
