+++ core/frontend/meta/canonical_url.js
@@ -1,5 +1,5 @@
 const _ = require('lodash');
-const urlUtils = require('../../lib/url-utils');
+const urlUtils = require('../../server/lib/url-utils');
 const getUrl = require('./url');
 
 function getCanonicalUrl(data) {
