+++ core/frontend/meta/image-dimensions.js
@@ -1,6 +1,6 @@
 var Promise = require('bluebird'),
     _ = require('lodash'),
-    imageLib = require('../../lib/image');
+    imageLib = require('../../server/lib/image');
 
 /**
  * Get Image dimensions
