+++ core/frontend/meta/index.js
@@ -1,7 +1,7 @@
 var Promise = require('bluebird'),
-    settingsCache = require('../../services/settings/cache'),
-    urlUtils = require('../../lib/url-utils'),
-    common = require('../../lib/common'),
+    settingsCache = require('../../server/services/settings/cache'),
+    urlUtils = require('../../server/lib/url-utils'),
+    common = require('../../server/lib/common'),
     getUrl = require('./url'),
     getImageDimensions = require('./image-dimensions'),
     getCanonicalUrl = require('./canonical_url'),
