+++ core/frontend/meta/rss_url.js
@@ -1,4 +1,4 @@
-const routingService = require('../../../frontend/services/routing');
+const routingService = require('../services/routing');
 
 function getRssUrl(data, absolute) {
     return routingService.registry.getRssUrl({
