+++ core/frontend/meta/structured_data.js
