+++ core/test/unit/data/meta/author_url_spec.js
@@ -2,7 +2,7 @@ const should = require('should'),
     sinon = require('sinon'),
     ObjectId = require('bson-objectid'),
     urlService = require('../../../../frontend/services/url'),
-    getAuthorUrl = require('../../../../server/data/meta/author_url');
+    getAuthorUrl = require('../../../../frontend/meta/author_url');
 
 describe('getAuthorUrl', function () {
     beforeEach(function () {
